Moved checks for string literal errors (unsupported escape characters, etc.) from binder to checker for performance reasons.
